move IOHeartbeatThread into its own file
Previously, the IOHeartbeatThread was located in DatabaseFeature.h/.cpp.
With the IOHeartbeatThread in its own file, several standard library
includes in DatabaseFeature.h can be avoided.
